feat: migrate to kord@0.7 and kordx@0.4 (SNAPSHOTS)

+++ src/main/kotlin/to/rxs/kommunity/commands/EvalCommand.kt
@@ -1,12 +1,12 @@
 package to.rxs.kommunity.commands
 
-import com.gitlab.kordlib.core.behavior.edit
-import com.gitlab.kordlib.kordx.commands.annotation.AutoWired
-import com.gitlab.kordlib.kordx.commands.annotation.ModuleName
-import com.gitlab.kordlib.kordx.commands.argument.text.StringArgument
-import com.gitlab.kordlib.kordx.commands.kord.model.respondEmbed
-import com.gitlab.kordlib.kordx.commands.kord.module.command
-import com.gitlab.kordlib.kordx.commands.model.command.invoke
+import dev.kord.core.behavior.edit
+import dev.kord.x.commands.annotation.AutoWired
+import dev.kord.x.commands.annotation.ModuleName
+import dev.kord.x.commands.argument.text.StringArgument
+import dev.kord.x.commands.kord.model.respondEmbed
+import dev.kord.x.commands.kord.module.command
+import dev.kord.x.commands.model.command.invoke
 import mu.KotlinLogging
 import java.time.Duration
 import java.time.Instant
@@ -19,7 +19,9 @@ private val log = KotlinLogging.logger {}
 @ModuleName("owner")
 fun evalCommand() = command("eval") {
     precondition {
-        if (kord.rest.application.getCurrentApplicationInfo().team?.members?.any { it.user.id.toLong() == author.id.longValue }!!.or(false)) {
+        if (kord.rest.application.getCurrentApplicationInfo().team?.members?.any { it.user.id == author.id }!!
+                .or(false)
+        ) {
             true
         } else {
             respond("This command can only be executed by an owner of the application!")
